Better highlighting of strings in coqide.
Strings were not highlighted out of comments. This would lead to funny
result, like certain strings (i.e. "(*") to cause all following code to be
highlighted as a string.

I've added strings in three different contexts (the same where comments
are highlighted). I think it's safe to do, I don't know if it's the best way,
though. In particular I don't know if it's the best way to highlight notations.

git-svn-id: svn+ssh://scm.gforge.inria.fr/svn/coq/trunk@15761 85f007b7-540e-0410-9357-904b9bb8a0f7